cleaned up files and created backup db
