--- all_horse_meta.csv
@@ -1,32 +1,0 @@
-,data.horse.bloodline,data.horse.breed_type,data.horse.breeding_counter,data.horse.color,data.horse.gen,data.horse.horse_type,data.horse.img_url,data.horse.inserted_at,data.horse.last_breeding_reset,data.horse.name,data.horse.nft_id,data.horse.race_statistic.first_place_finishes,data.horse.race_statistic.number_of_races,data.horse.race_statistic.second_place_finishes,data.horse.race_statistic.third_place_finishes,data.horse.race_statistic.win_rate
-0,Buterin,exclusive,3,Dark Byzantium,Z10,Stallion,https://img.zed.run/horses/5D3954.png,2021-11-26T20:32:39,2022-12-23T20:28:07Z,Scrunter S Thompson,154936,11,101,7,10,10.89
-0,Buterin,elite,1,Banana Mania,Z23,Filly,https://img.zed.run/horses/FAE7B5.png,2022-04-15T08:19:30,2022-12-23T08:15:00Z,Considerable Threat,393229,1,75,5,9,1.33
-0,Nakamoto,exclusive,3,Fandango,Z5,Colt,https://img.zed.run/horses/B53389.png,2022-04-15T08:36:23,2022-12-23T08:31:54Z,Nakuza,393233,6,131,15,12,4.58
-0,Finney,legendary,3,Air Superiority Blue,Z7,Stallion,https://img.zed.run/horses/72A0C1.png,2021-10-17T06:37:21,2023-01-08T06:30:43Z,Glorified Liberty,131092,5,65,7,8,7.69
-0,Nakamoto,exclusive,3,Battle Horse Gray,Z6,Colt,https://img.zed.run/horses/848482.png,2021-10-17T06:37:43,2023-01-08T06:31:09Z,Glorified Ecstasy,131093,12,151,9,15,7.95
-0,Buterin,elite,3,Eminence,Z16,Colt,https://img.zed.run/horses/6C3082.png,2022-04-15T08:51:24,2022-12-23T08:46:42Z,System America,393237,0,11,0,1,0.0
-0,Nakamoto,genesis,1,Absolute Zero,Z1,Mare,https://img.zed.run/horses/0048BA.png,2019-01-31T03:49:18,2022-12-29T03:48:43Z,Troubled Paradise,27,225,1819,254,236,12.37
-0,Buterin,exclusive,1,Feldgrau,Z28,Filly,https://img.zed.run/horses/4D5D53.png,2021-10-17T06:51:45,2023-01-08T06:45:07Z,Boron Metalloid,131100,0,10,0,0,0.0
-0,Buterin,exclusive,1,Fawn,Z24,Filly,https://img.zed.run/horses/E5AA70.png,2021-10-17T06:53:12,2023-01-08T06:46:33Z,Be Water My Friend,131102,0,13,1,0,0.0
-0,Buterin,exclusive,3,Dark Lava,Z27,Colt,https://img.zed.run/horses/483C32.png,2021-10-17T06:55:42,2023-01-08T06:49:05Z,Be Kind,131104,18,142,8,3,12.68
-0,Buterin,elite,1,Electric Violet,Z42,Filly,https://img.zed.run/horses/8F00FF.png,2021-10-17T06:58:30,2023-01-08T06:51:55Z,Be Formless,131105,0,1,0,0,0.0
-0,Buterin,exclusive,1,Golden Rod,Z20,Mare,https://img.zed.run/horses/EEE8AA.png,2022-04-15T09:52:00,2022-12-23T09:47:31Z,Tough Name,393248,16,107,7,13,14.95
-0,Nakamoto,elite,1,Egyptian Blue,Z6,Mare,https://img.zed.run/horses/1034A6.png,2022-04-15T10:05:06,2022-12-23T10:00:35Z,Ats Us Nai,393260,17,156,11,12,10.9
-0,Finney,exclusive,1,Dartmouth Green,Z12,Mare,https://img.zed.run/horses/00703C.png,2021-10-17T08:58:30,2023-01-08T08:52:33Z,Beast Within,131127,4,210,17,21,1.9
-0,Finney,exclusive,3,Cerulean Frost,Z15,Stallion,https://img.zed.run/horses/6D9BC3.png,2021-10-17T09:04:26,2023-01-08T08:57:49Z,Dig Through Time,131128,156,1432,99,87,10.89
-0,Szabo,legendary,1,Oval Orchid,Z4,Mare,https://img.zed.run/horses/DA70D6.png,2022-04-15T10:18:42,2022-12-23T10:14:15Z,Nefarious Child,393271,15,274,19,40,5.47
-0,Finney,elite,1,Bondi Blue,Z20,Filly,https://img.zed.run/horses/0095B6.png,2022-04-15T10:41:55,2022-12-23T10:35:20Z,Sapphire Shine,393279,0,47,2,6,0.0
-0,Szabo,elite,1,Amethyst,Z16,Filly,https://img.zed.run/horses/9966CC.png,2022-04-15T10:42:44,2022-12-23T10:36:36Z,Amethyst Shine,393280,5,300,21,37,1.67
-0,Finney,exclusive,1,Caribbean Green,Z25,Mare,https://img.zed.run/horses/00CC99.png,2022-04-15T10:46:13,2022-12-23T10:41:42Z,Jade Shine,393281,25,347,40,37,7.2
-0,Nakamoto,elite,3,Black Chocolate,Z10,Colt,https://img.zed.run/horses/1B1811.png,2022-04-15T10:54:29,2022-12-23T10:49:58Z,Come Again For Big Fudge,393282,12,297,10,17,4.04
-0,Nakamoto,elite,3,Shadow Fighter,Z12,Colt,https://img.zed.run/horses/BFAFB2.png,2022-04-15T10:55:08,2022-12-23T10:50:40Z,Lurking In The Shadows,393284,3,55,4,6,5.45
-0,Nakamoto,genesis,3,Slate Gray,Z1,Stallion,https://img.zed.run/horses/708090.png,2019-02-07T09:20:56,2023-01-05T09:20:41Z,Mass Anchor,68,5,38,5,12,13.16
-0,Buterin,exclusive,3,Alizarin Crimson,Z21,Colt,https://img.zed.run/horses/E32636.png,2022-04-15T11:03:22,2022-12-23T10:58:16Z,Gotta Start Somewhere,393288,5,103,13,5,4.85
-0,Nakamoto,genesis,1,Slate Gray,Z1,Mare,https://img.zed.run/horses/708090.png,2019-02-07T11:15:36,2023-01-05T11:15:31Z,Hater,75,4,20,5,2,20.0
-0,Szabo,legendary,1,Azure,Z5,Mare,https://img.zed.run/horses/007FFF.png,2021-10-17T11:27:26,2023-01-08T11:21:06Z,Diamond Cut,131156,29,367,37,40,7.9
-0,Szabo,legendary,1,Absolute Zero,Z6,Mare,https://img.zed.run/horses/0048BA.png,2022-04-15T12:12:08,2022-12-23T12:07:42Z,Agent Know How,393302,8,120,6,7,6.67
-0,Finney,exclusive,1,Feldgrau,Z9,Mare,https://img.zed.run/horses/4D5D53.png,2022-04-15T12:33:13,2022-12-23T12:26:36Z,High Gin-Shiki,393303,90,983,66,73,9.16
-0,Nakamoto,genesis,1,Slate Gray,Z1,Mare,https://img.zed.run/horses/708090.png,2019-02-15T19:45:22,2023-01-13T19:45:15Z,Two Birds One Stone,89,5,23,3,5,21.74
-0,Finney,exclusive,3,Midnight Black,Z16,Stallion,https://img.zed.run/horses/000000.png,2022-04-15T12:37:05,2022-12-23T12:32:30Z,Midnight Mash,393305,19,138,8,9,13.77
-0,Nakamoto,genesis,3,Champagne Papi,Z1,Stallion,https://img.zed.run/horses/F1DDCF.png,2019-02-16T13:10:03,2023-01-14T13:09:35Z,Nine Nine Two,92,28,276,37,40,10.14
-0,Finney,legendary,1,Egyptian Blue,Z12,Filly,https://img.zed.run/horses/1034A6.png,2022-04-15T13:11:55,2022-12-23T13:07:24Z,Drunken Monkey,393315,3,30,2,0,10.0
