+++ assets/example_data/1_horse_meta_data.csv
@@ -1,0 +1,2 @@
+,data.horse.bloodline,data.horse.breed_type,data.horse.color,data.horse.gen,data.horse.horse_type,data.horse.img_url,data.horse.inserted_at,data.horse.name,data.horse.nft_id,data.horse.parents,data.horse.race_statistic.first_place_finishes,data.horse.race_statistic.free_win_rate,data.horse.race_statistic.number_of_free_races,data.horse.race_statistic.number_of_paid_races,data.horse.race_statistic.number_of_races,data.horse.race_statistic.paid_win_rate,data.horse.race_statistic.second_place_finishes,data.horse.race_statistic.third_place_finishes,data.horse.race_statistic.win_rate,data.horse.super_coat
+0,Buterin,genesis,Powder Blue,Z10,Stallion,https://img.zed.run/horses/B0E0E6.png,2020-10-09T01:05:13,Midnightmane,8919,[],199,11.48,1167,670,1837,9.7,173,147,10.83,False
